fix: fully static
